+++ tests/Fake/ReverseShortcode.php
@@ -1,0 +1,12 @@
+<?php
+namespace Thunder\Shortcode\Tests\Fake;
+
+use Thunder\Shortcode\Shortcode;
+
+class ReverseShortcode
+    {
+    public function __invoke(Shortcode $s)
+        {
+        return strrev($s->getContent());
+        }
+    }