+++ tests/ProcessorTest.php
@@ -7,6 +7,7 @@ use Thunder\Shortcode\Processor;
 use Thunder\Shortcode\Shortcode;
 use Thunder\Shortcode\Shortcode\ProcessedShortcode;
 use Thunder\Shortcode\Tests\Fake\HtmlShortcode;
+use Thunder\Shortcode\Tests\Fake\ReverseShortcode;
 
 /**
  * @author Tomasz Kowalczyk <tomasz@kowalczyk.cc>
@@ -21,6 +22,7 @@ final class ProcessorTest extends \PHPUnit_Framework_TestCase
             ->addHandler('name', function(Shortcode $s) { return $s->getName(); })
             ->addHandler('content', function(Shortcode $s) { return $s->getContent(); })
             ->addHandler('html', new HtmlShortcode())
+            ->addHandler('reverse', new ReverseShortcode())
             ->addHandlerAlias('c', 'content')
             ->addHandlerAlias('n', 'name');
 
@@ -51,6 +53,7 @@ final class ProcessorTest extends \PHPUnit_Framework_TestCase
             array('x [content]a-[c]v[/c]-b[/content] y', 'x a-v-b y'),
             array('x [html b]bold[/html] y [html code]code[/html] z', 'x <b>bold</b> y <code>code</code> z'),
             array('x [html]bold[/html] z', 'x [html]bold[/html] z'),
+            array('x [reverse]abc xyz[/reverse] z', 'x zyx cba z'),
             );
         }